test: re-organize CustomField ITs
