+++ vaadin-custom-field-flow-parent/vaadin-custom-field-flow-integration-tests/src/main/java/com/vaadin/flow/component/customfield/BasicPage.java
@@ -1,15 +1,14 @@
-package com.vaadin.flow.component.customfield.examples;
+package com.vaadin.flow.component.customfield;
 
-import com.vaadin.flow.component.customfield.CustomField;
 import com.vaadin.flow.component.html.Div;
 import com.vaadin.flow.component.html.NativeButton;
 import com.vaadin.flow.component.textfield.TextField;
 import com.vaadin.flow.router.Route;
 
 @Route(value = "vaadin-custom-field")
-public class Home extends Div {
+public class BasicPage extends Div {
 
-    public Home() {
+    public BasicPage() {
         final Div result = new Div();
         result.setId("result");
         MyField customField = new MyField();
