--- vaadin-custom-field-flow-parent/vaadin-custom-field-flow-integration-tests/src/main/java/com/vaadin/flow/component/customfield/examples/NameField.java
@@ -1,36 +1,0 @@
-package com.vaadin.flow.component.customfield.examples;
-
-import com.vaadin.flow.component.customfield.CustomField;
-import com.vaadin.flow.component.html.Div;
-import com.vaadin.flow.component.textfield.TextField;
-
-public class NameField extends CustomField<String> {
-    private final TextField firstName = new TextField();
-    private final TextField lastName = new TextField();
-
-    NameField() {
-        setLabel("Phone number");
-        setHelperText("Your full first and last names");
-        firstName.setMinLength(2);
-        firstName.getStyle().set("width", "7em");
-        lastName.getStyle().set("width", "7em");
-
-        Div layout = new Div();
-        layout.add(firstName, lastName);
-
-        add(layout);
-    }
-
-    @Override
-    protected String generateModelValue() {
-        return firstName.getValue() + " " + lastName.getValue();
-    }
-
-    @Override
-    protected void setPresentationValue(String newPresentationValue) {
-        if (newPresentationValue == null) {
-            firstName.clear();
-            lastName.clear();
-        }
-    }
-}
