+++ vaadin-custom-field-flow-parent/vaadin-custom-field-flow-integration-tests/src/test/java/com/vaadin/flow/component/customfield/test/ErrorIT.java
@@ -11,9 +11,7 @@ import com.vaadin.flow.component.customfield.testbench.CustomFieldElement;
 public class ErrorIT extends AbstractParallelTest {
     @Before
     public void init() {
-        String url = getBaseURL().replace(super.getBaseURL(),
-                super.getBaseURL() + "/vaadin-custom-field") + "/error";
-        getDriver().get(url);
+        getDriver().get(getBaseURL() + "/vaadin-custom-field/error");
     }
 
     @Test
